Ensure that blocklist contains no duplicate IDs.
Since blocks are identified by their "id" field, weird behaviour
can occur when there are duplicates. This change ensures that all
IDs are unique after the blocklist is populated.

By default, this is achieved by mangling the IDs of any duplicated
blocks. On the other hand, if the "purge_duplicated_ids" flag of the
filehandle is set then duplicated blocks will be discarded.