Fix a memory leak.
The memory allocated in lines 108-109 would never be freed if h->mmap is
true. I've read through the code called by 'sdf_read_data' on line 111,
and I think these lines are redundant. This should be double-checked,
however.

Also, 'if (b && !b->data)' in line 104 guarantees that the 'free'
statement in line 107 will never be executed, so line 107 is redundant.

This commit can be fixed up with the previous one (ie 'Fix some memory
leaks') once the reviewer is satisfied it's correct.