Search for extension library using SDF_EXTENSION_PATH
The SDF_EXTENSION_PATH environment variable should be a string
separated by colon, semicolon or comma. Each substring is examined
in turn. If it is a file then we attempt to load it as a dynamic
library. If it is a directory then we attempt to load a file named
"sdf_extension.so" or "libsdf_extension.so" in that directory. The
routine exits once a valid library is found.